E-Government revision
First I revised the text and checked the information. Overall the spelling and grammar were good.  Then I reorganized some parts, I added textual additions and created sections and I added the main benefits of this new type of government.  I added a link of an image and a link of a YouTube video about E-Government